+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Use `defaultbrowser` to set Firefox as the default on macOS
+- Added `defaultbrowser` overlay
 - Refactored NixOS and `nix-darwin` modules in `flake.nix` into `base` module
 - Fixed hostname not being set on macOS
 - Changed `ios` module to be Linux only